NOBUG: Organize component files
