--- src/gatsby/src/components/HTMLArea.js
@@ -1,13 +1,0 @@
-import React from "react"
-import { usePreRenderVideo } from "../utils/usePreRenderVideo"
-
-export default function HTMLArea(props) {
-  const { htmlContent } = usePreRenderVideo(props.children)
-
-  if (!props.isVisible) {
-    return null;
-  }
-  return (
-    <div dangerouslySetInnerHTML={{ __html: htmlContent || props.children}}/>
-  )
-}
