--- src/gatsby/src/components/advisories/advisoryAlerts.js
@@ -1,3 +1,0 @@
-
-// TODO create helper function to store common alert data (e.g. colors)
-// to be used by other advisory components
