--- src/gatsby/src/components/media.js
@@ -1,11 +1,0 @@
-import React from "react"
-
-export default function Media(props)
-{
-  if (!props.isVisible) {
-    return null;
-  }
-    return (
-    <img src={props.url} alt="Example" />
-  )
-}
