+++ src/gatsby/src/components/park/safetyInfo.js
@@ -1,5 +1,5 @@
 import React from "react"
-import HtmlContent from "./htmlContent"
+import HtmlContent from "../htmlContent"
 
 export default function SafetyInfo({ safetyInfo }) {
   return (
